nrf52840: fix 32-bit TIMER compare wrap overflow (boot IRQ storm)
timer_schedule_next_compare computed the wrapping distance to a compare
value as `uint32_t ticks = (cc - now) & mask; if (ticks == 0) ticks =
mask + 1;` — with BITMODE=32 the mask is 0xFFFFFFFF, so mask+1
overflowed to 0 and the dc<1 clamp re-armed the compare event ONE CPU
CYCLE ahead, forever.

Current Contiki-NG's nrf52840 rtimer-arch starts TIMER0 exactly into
this hole: 32-bit mode, COMPARE0 interrupt enabled at init, TASKS_START
with CC[0] still 0 == counter.  Result: a boot-time COMPARE0 interrupt
storm that pinned the CPU in the TIMER0 ISR at full speed — TSCH slot
operation never ran (coordinator sent zero EBs), and the simulation
crawled at ~0.02x realtime.  The shipped CSMA firmware predates the
rtimer-arch that enables the interrupt at init, which is why the storm
never showed before.

Do the equal-means-full-wrap arithmetic in 64-bit.  With this, 2-node
nrf52840-dk TSCH (fresh 6tisch/simple-node build) associates in 0.14 s
and holds sync for 180 s with zero desyncs (EB-based resync; unicast
EACK timing is a separate remaining item).  Sim speed on that firmware:
0.02x -> 1.4x realtime.

Regressions green: nrf52840-dk RPL-UDP, Zephyr echo (110 replies),
nrf54l15 FLPR, arm-correctness 153/153.

+++ src/arm/nrf52840_soc.c
@@ -1916,9 +1916,17 @@ static void timer_schedule_next_compare(nrf_timer_state_t *t) {
                     has_listener = true;
         }
         if (!has_listener) continue;
-        uint32_t ticks = (t->cc[i] - now) & mask;
-        if (ticks == 0) ticks = mask + 1;
-        int64_t dc = (int64_t)ticks * (int64_t)cpu->cpu_freq_hz / (int64_t)tick_hz;
+        /* Distance to the compare value on the wrapping counter.  Do the
+         * "equal means a full wrap" case in 64-bit: with BITMODE=32 the
+         * mask is 0xFFFFFFFF and a uint32 mask+1 overflowed to 0, which
+         * the dc<1 clamp turned into a compare event every single CPU
+         * cycle.  Contiki's rtimer-arch starts TIMER0 (32-bit, CC0=0,
+         * COMPARE0 int enabled) with counter==CC0, hitting exactly this:
+         * a boot-time interrupt storm that pinned the CPU in the TIMER0
+         * ISR so TSCH slot operation never ran. */
+        uint64_t ticks = (uint64_t)((t->cc[i] - now) & mask);
+        if (ticks == 0) ticks = (uint64_t)mask + 1;
+        int64_t dc = (int64_t)(ticks * (uint64_t)cpu->cpu_freq_hz / tick_hz);
         if (dc < 1) dc = 1;  /* avoid 0-delay infinite reschedule */
         if (dc < best_delta) { best_delta = dc; best_i = i; }
     }
